Moved canopy type into own file
